persimmon_chatbot/requirements.txt

streamlit==1.37.0
pdfplumber==0.11.5
google-generativeai==0.8.3
faster-whisper==1.0.3
streamlit-webrtc==0.47.7
pyav==12.3.0
numpy==1.26.4
soundfile==0.12.1
langgraph==0.2.14
torch==2.4.1
